# api/src/main/resources/public/swagger/api.yaml
openapi: 3.0.2
servers:
  - url: http://localhost:8080/
info:
  description: This is a OpenApi specification for MyHome backend service.
  version: 2.0.0
  title: Swagger MyHome - OpenAPI 3.0
...
paths:
...
  /communities/{communityId}/houses:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Houses
      description: List all houses of the community given a community id
      operationId: listCommunityHouses
      parameters:
        - in: path
          name: communityId
          schema:
            type: string
          required: true
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: If community exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetHouseDetailsResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetHouseDetailsResponse'
        '404':
          description: If params are invalid
    post:
      tags:
        - Houses
      description: Add a new house to the community given a community id
      operationId: addCommunityHouses
      parameters:
        - in: path
          name: communityId
          schema:
            type: string
          required: true
      requestBody:
        description: AddCommunityHouseRequest aggregate fields
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCommunityHouseRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/AddCommunityHouseRequest'
      responses:
        '204':
          description: If houses were added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddCommunityHouseResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/AddCommunityHouseResponse'
        '400':
          description: If params are invalid
  /communities/{communityId}/houses/{houseId}:
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - Houses
      description: Remove of house from the community given a community id and a house id
      operationId: removeCommunityHouse
      parameters:
        - in: path
          name: communityId
          schema:
            type: string
          required: true
        - in: path
          name: houseId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: If house was removed
        '400':
          description: If params are invalid
  /houses:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Houses
      description: List all houses of the community given a community id
      operationId: listAllHouses
      parameters:
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: If community exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetHouseDetailsResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetHouseDetailsResponse'
  /houses/{houseId}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Houses
      description: Get house detail by a given ID
      operationId: getHouseDetails
      parameters:
        - in: path
          name: houseId
          schema:
            type: string
          required: true
          description: ID of the house to get
      responses:
        '200':
          description: If house present
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetHouseDetailsResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/GetHouseDetailsResponse'
        '404':
          description: If params are invalid
  /houses/{houseId}/members:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Houses
      description: List all members of the house given a house id
      operationId: listAllMembersOfHouse
      parameters:
        - in: path
          name: houseId
          schema:
            type: string
          required: true
        - in: query
          name: pageable
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: If house present
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListHouseMembersResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ListHouseMembersResponse'
        '404':
          description: If params are invalid
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Houses
      description: Add new members to the house given a house id. Responds with member id which were added
      operationId: addHouseMembers
      parameters:
        - in: path
          name: houseId
          schema:
            type: string
          required: true
      requestBody:
        description: AddHouseMemberRequest aggregate fields
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddHouseMemberRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/AddHouseMemberRequest'
      responses:
        '201':
          description: If members were added to house
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddHouseMemberResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/AddHouseMemberResponse'
        '404':
          description: If parameters are invalid
  /houses/{houseId}/members/{memberId}:
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - Houses
      description: Deletion of member associated with a house
      operationId: deleteHouseMember
      parameters:
        - in: path
          name: houseId
          schema:
            type: string
          required: true
        - in: path
          name: memberId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: If house member was removed from house
        '400':
          description: If params are invalid
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pageable:
      description: Pageable query parameters
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
...
    GetHouseDetailsResponse:
      type: object
      required:
        - houses
      properties:
        houses:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/GetHouseDetailsResponseCommunityHouse'
    GetHouseDetailsResponseCommunityHouse:
      type: object
      properties:
        houseId:
          type: string
        name:
          type: string
    AddCommunityHouseRequest:
      type: object
      required:
        - houses
      properties:
        houses:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/CommunityHouseName'
    CommunityHouseName:
      type: object
      properties:
        name:
          type: string
    AddCommunityHouseResponse:
      type: object
      required:
        - houses
      properties:
        houses:
          type: array
          uniqueItems: true
          items:
            type: string
    HouseMember:
      type: object
      required:
        - memberId
        - name
      properties:
        memberId:
          type: string
        name:
          type: string
    ListHouseMembersResponse:
      type: object
      required:
        - members
      properties:
        members:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/HouseMember'
    HouseMemberDto:
      type: object
      required:
        - memberId
        - name
      properties:
        id:
          type: integer
          format: int64
        memberId:
          type: string
        name:
          type: string
    AddHouseMemberRequest:
      type: object
      required:
        - members
      properties:
        members:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/HouseMemberDto'
    AddHouseMemberResponse:
      type: object
      required:
        - members
      properties:
        members:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/HouseMember'
...